Timer Class update
Changes made to Timer class and ojbect to account for more precise
timming of events.  Ideally reaching for timming down in the microsecond
range in order to accurately control the number of times that the RPi is
requesting data from the Arduino.

Idealy this request rate should be around 1MHz or slightly faster.

Update also includes changes to the pom file to account for duplicate
class deffinitions in junit and junit's dependency on hamcrest-core.